+++ src/Vulkan/VkBufferView.hpp
@@ -35,8 +35,8 @@ public:
 	}
 
 	void *getPointer() const;
-	uint32_t getElementCount() const { return range / Format(format).bytes(); }
-	uint32_t getRangeInBytes() const { return range; }
+	uint32_t getElementCount() const { return static_cast<uint32_t>(range / Format(format).bytes()); }
+	uint32_t getRangeInBytes() const { return static_cast<uint32_t>(range); }
 	VkFormat getFormat() const { return format; }
 
 	const uint32_t id = ImageView::nextID++;	// ID space for sampling function cache, shared with imageviews
